requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
pandas>=2.2.2
pydantic>=2.7.0
rich>=13.7.1
streamlit>=1.36.0
PyJWT>=2.8.0
plotly>=5.22.0
websocket-client>=1.8.0
streamlit-aggrid>=0.3.4
